refactor: extract editor dock panels
