+++ packages/editor/src/panels/PanelShell.tsx
@@ -1,0 +1,45 @@
+import type React from 'react';
+import { v } from '../theme/tokens';
+
+export function PanelShell({
+  title,
+  children,
+  style,
+}: {
+  title: string;
+  children: React.ReactNode;
+  style?: React.CSSProperties;
+}) {
+  return (
+    <section
+      style={{
+        minHeight: 0,
+        display: 'flex',
+        flexDirection: 'column',
+        overflow: 'hidden',
+        ...style,
+      }}
+    >
+      <div
+        style={{
+          height: 30,
+          display: 'flex',
+          alignItems: 'center',
+          padding: '0 10px',
+          borderBottom: `1px solid ${v('--elucim-editor-border-subtle')}`,
+          color: v('--elucim-editor-text-secondary'),
+          fontSize: 10,
+          fontWeight: 700,
+          letterSpacing: 0.8,
+          textTransform: 'uppercase',
+          flexShrink: 0,
+        }}
+      >
+        {title}
+      </div>
+      <div style={{ flex: 1, minHeight: 0, overflow: 'auto' }}>
+        {children}
+      </div>
+    </section>
+  );
+}
